Add shared state and use nested pages for pathapp
