+++ client/app.vue
@@ -9,15 +9,6 @@
 </template>
 
 <style lang='scss'>
-/* NavBar is fixed to the top */
-.main-nav {
-    position: fixed;
-    top: 0;
-    left: 0;
-    z-index: 9999;
-
-    width: 100%;
-}
 /* Body is under the NavBar */
 .nav-bumper {
     margin-top: $navbar-height;
