+++ client/pages/login.vue
@@ -1,4 +1,4 @@
-<script setup lang="ts">
+<script setup>
 useHead({
     title: 'Login'
 })
