+++ client/pages/pathapp/index.vue
@@ -1,3 +1,8 @@
 <template>
-    <div>This is the task picker.</div>
+    <div class="container p-5">
+        <div class="title">Pick a task</div>
+        <div class=''>
+            <TaskPickerList />
+        </div>
+    </div>
 </template>
